feat(parse): type check, name to UnaryExpr only
With type-checking, different types of unary expressions can be
differentiated without defining new grammar rules. This is, well, nice.
But it does make our parser not 100% context-free.

To be fair, the grammar may be language free still. But, it's hard for
the parser to be context-free. Say, later on, we add variable
declaration. If the parser is context-free, how should we bind
variables? Well, probably by running a "linker" parser that follows the
parse tree and substitutes variable references. I'm probably not doing a
one-pass parser. So, that could be a solution.

Still, there are a lot of cases, a variable needs to be declared before
use. Say, in a block. I can, of course, still breeze through the parse
tree with the "linker," but it's faster to do it in this first phase.

Well, yknow what, maybe I will actually go with a context-free grammar
parser, then a definitely-non-context-free linker/type-checker.

Commit-by: Huy Nguyen