feat(parse): store type tag as struct member
Trades quite a bit of memory for less need to kind-of backtrack (since,
each time type_tag is called before, it usually just forwards that
request to a child node). A type_tag may contain a string, so it's
pretty heavy.

Maybe Imma make it a vector of strings instead, when I get to the part
of namespace resolution. Say, how would one differentiate std::string
from my_crate::string? Easiest way is to, well, store the double colon
as the string as well. Which, actually may be faster than a vector of
strings, since it's now a continuous array, instead of an array of
pointers.

Commit-by: Huy Nguyen