wip(parse): add typetag to all expressions.
Statements really only have one type, that is, the unit type, or the
void type. However you would call it.

Also, I just realized that this mechanism of type-checking makes it very
hard, or even, impossible, to write a multi-pass parser that resolves
typing later on. Well, let's aim for the simplest possible.

Commit-by: Huy Nguyen