fix(parse): ProcParams pop tokens way too eagerly.
So, I allow empty ProcParams. Which means, the first check should just
peek into the token list with front, *not* pop_front. I can only
confidently pop if I'm sure that the type of token I'm looking for is
correct. Similarly with the while-loop that checks for a comma.

The checking-for-Colon part should still pop_front eagerly, because if
it's not a Colon, it's an error anyways.

Also, I decided I will allow a sole comma "," as a valid ProcParams. It
will look a bit bizzare (eg, "def procname = proc (,) {};" is a valid
formation). I will need to update the grammar rule again for that.

Commit-by: Huy Nguyen