feat(parse): new_from for IfExpr
First, IfExpr cannot really fail with Err(None). If it runs out of
tokens, it actually just fails, unlike some other expressions.
Second, maybe some future constructs will only need a concrete IfExpr.
The simplest I can think of is making an IfStmt that pretty much just
wraps around an IfExpr.
Why? With the current implementation, to fit an IfExpr inside a block,
it has to be a statement, and u need a semicolon for that.
If I do make an IfStmt, I will need to update the grammar of
ExprSemicolonStmt.

Signed-off-by: Huy Nguyen <huy55465@gmail.com>