fix(parse): LetStmt owns a Rc<String>, ScopedBlock
Because I will need ScopedBlock to be able to refer to the identifier
name held by LetStmt.
Later on, I may need to extract retrieving this Rc<String> to
DeclStmtImpl's trait.
The more I'm working on this, the more I don't like OOP in Rust.
Also, I did some testing on ScopedBlock.

Signed-off-by: Huy Nguyen <huy55465@gmail.com>