BREAKING CHANGE: remove some 2-char symbols
And replace them with keywords.
This simplifies the tokenizer state machine by a fair bit. Are these new
keywords more readable? Perhaps, and perhaps not. But they convey the
meaning, don't they?

The 3 remaining 2-char symbols are <= >= and //
The former two are standard mathematical symbols. They are, IMO, the
easiest to recognize already.
The latter; I may consider using the hash symbol. It eliminates yet
another state.

Signed-off-by: Huy Nguyen <huy55465@gmail.com>