fix(tokenize): integer + dot
The case the dot is not followed by a number or space characters (our
language only supports ' ', '\t' and '\n') was broken.
Because it will match the _ case, which is an auto-error.
This fixes it by separating the dot-case and the rest-case.

And, add 2 more test cases for this situation.

Signed-off-by: Huy Nguyen <huy55465@gmail.com>